% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcavatedTtest.R
\name{ExcavatedTtest}
\alias{ExcavatedTtest}
\title{Run a T-test on excavated and non-excavated mussels Note: filter out the non-excavated and excavated dataframes. # Nonexcavated_mussels <- filter(SH16_Mussels, Excavated == "N") \%>\%}
\usage{
ExcavatedTtest(Nonexcavated_list, Excavated_list)
}
\arguments{
\item{Nonexcavated_list}{A list of nonexcavated mussel metrics}

\item{Excavated_list}{A list of excavated mussel variables
@return Dataframe A dataframe of the T-test model
@usage (Nonexcavated_list, Excavated_list)}
}
\description{
Run a T-test on excavated and non-excavated mussels Note: filter out the non-excavated and excavated dataframes. # Nonexcavated_mussels <- filter(SH16_Mussels, Excavated == "N") \%>\%
}
