% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModelWidth.R
\name{LinearModelLength}
\alias{LinearModelLength}
\title{Create a linear model based on length and width from dataframe}
\usage{
LinearModelLength(Mussel_data, Variable_1, Variable_2)
}
\arguments{
\item{Mussel_data}{A dataframe of mussel data}

\item{Variable_1}{A column of data within mussel data (ie. Mussel_data$variable_1)}

\item{Variable_2}{A column of data within mussel data (ie. Mussel_data$variable_2)}
}
\value{
A Linear Model
}
\description{
Create a linear model based on length and width from dataframe
}
