% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlotDepthLength.R
\name{ScatterPlotDepthLength}
\alias{ScatterPlotDepthLength}
\title{Make a plot based on depth and length of mussels.}
\usage{
ScatterPlotDepthLength(Mussel_data, Depth_Variable, Length_Variable)
}
\arguments{
\item{Mussel_data}{A dataframe of mussel metrics}

\item{Depth_variable}{A subsetted column of depth variables}

\item{Length_variable}{A subsetted column of length variables}

\item{X_axis}{label A label text for the x-axis}

\item{Y_axis_label}{A label text for the y_axis}

\item{Title}{A Title text}
}
\value{
A Plot
}
\description{
Make a plot based on depth and length of mussels.
}
