% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlotDepthWidth.R
\name{ScatterPlotDepthWidth}
\alias{ScatterPlotDepthWidth}
\title{Make a scatterplot based on depth and width of mussels}
\usage{
(Mussel_data, Depth_Variable, Width_variable, Y_axis_label, X_axis_label, Title)
}
\arguments{
\item{Mussel_data}{A dataframe of mussel metrics}

\item{Depth_variable}{A subsetted column of depth variables}

\item{Width_variable}{A subsetted column of width variables}

\item{X_axis}{label A label text for the x-axis}

\item{Y_axis_label}{A label text for the y_axis}

\item{Title}{A Title text}
}
\value{
A Plot
}
\description{
Make a scatterplot based on depth and width of mussels
}
